Find dependencies before definining targets
Previously we were defining our exported targets before finding their
dependencies. As a consequence, when the exported configuration was found by an
external project, it failed with a message saying that targets had been found,
but not some of its dependencies, and the overall cmake configuration failed.
Finding dependencies before defining targets successfully fixes this problem.